+++ lib/spout/commands/project_generator.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
 require 'colorize'
-require 'date'
-require 'erb'
 require 'fileutils'
 
-TEMPLATES_DIRECTORY = File.expand_path('../../templates', __FILE__)
+require 'spout/helpers/framework'
 
 module Spout
   module Commands
     # Generates folder and file structure for a new spout data dictionary.
     class ProjectGenerator
+      include Spout::Helpers::Framework
+
       def initialize(argv)
         generate_folder_structure!(argv)
       end
@@ -53,33 +53,6 @@ EOT
         puts '         run'.colorize(:green) + '  bundle install'.colorize(:light_cyan)
         Dir.chdir(@full_path)
         system 'bundle install'
-      end
-
-      private
-
-      def copy_file(template_file, file_name = '')
-        file_name = template_file if file_name == ''
-        file_path = File.join(@full_path, file_name)
-        template_file_path = File.join(TEMPLATES_DIRECTORY, template_file)
-        puts '      create'.colorize(:green) + "  #{file_name}"
-        FileUtils.copy(template_file_path, file_path)
-      end
-
-      def evaluate_file(template_file, file_name)
-        template_file_path = File.join(TEMPLATES_DIRECTORY, template_file)
-        template = ERB.new(File.read(template_file_path))
-        file_path = File.join(@full_path, file_name)
-        file_out = File.new(file_path, 'w')
-        file_out.syswrite(template.result(binding))
-        puts '      create'.colorize(:green) + "  #{file_name}"
-      ensure
-        file_out.close if file_out
-      end
-
-      def directory(directory_name)
-        directory_path = File.join(@full_path, directory_name)
-        puts '      create'.colorize(:green) + "  #{directory_name}"
-        FileUtils.mkpath(directory_path)
       end
     end
   end